+++ src/main/java/net/cupellation/init/RenderInit.java
@@ -23,12 +23,20 @@ public class RenderInit {
         BlockEntityRendererFactories.register(BlockInit.CASTING_BASIN_ENTITY, CastingBasinRenderer::new);
         BlockEntityRendererFactories.register(BlockInit.SMELTER_FAUCET_ENTITY, SmelterFaucetRenderer::new);
         BlockEntityRendererFactories.register(BlockInit.CASTING_TABLE_ENTITY, CastingTableRenderer::new);
+
         BlockRenderLayerMap.INSTANCE.putBlock(BlockInit.DEEPSLATE_BRICK_SMELTER, RenderLayer.getCutout());
         BlockRenderLayerMap.INSTANCE.putBlock(BlockInit.DEEPSLATE_BRICK_GLASS, RenderLayer.getCutout());
         BlockRenderLayerMap.INSTANCE.putBlock(BlockInit.DEEPSLATE_BRICK_CASTING_BASIN, RenderLayer.getCutout());
         BlockRenderLayerMap.INSTANCE.putBlock(BlockInit.DEEPSLATE_BRICK_CASTING_TABLE, RenderLayer.getCutout());
         BlockRenderLayerMap.INSTANCE.putBlock(BlockInit.DEEPSLATE_BRICK_LEVER, RenderLayer.getCutout());
         BlockRenderLayerMap.INSTANCE.putBlock(BlockInit.DEEPSLATE_BRICK_FAUCET, RenderLayer.getCutoutMipped());
+
+        BlockRenderLayerMap.INSTANCE.putBlock(BlockInit.RED_NETHER_BRICK_SMELTER, RenderLayer.getCutout());
+        BlockRenderLayerMap.INSTANCE.putBlock(BlockInit.RED_NETHER_BRICK_GLASS, RenderLayer.getCutout());
+        BlockRenderLayerMap.INSTANCE.putBlock(BlockInit.RED_NETHER_BRICK_CASTING_BASIN, RenderLayer.getCutout());
+        BlockRenderLayerMap.INSTANCE.putBlock(BlockInit.RED_NETHER_BRICK_CASTING_TABLE, RenderLayer.getCutout());
+        BlockRenderLayerMap.INSTANCE.putBlock(BlockInit.RED_NETHER_BRICK_LEVER, RenderLayer.getCutout());
+        BlockRenderLayerMap.INSTANCE.putBlock(BlockInit.RED_NETHER_BRICK_FAUCET, RenderLayer.getCutoutMipped());
 
         HandledScreens.register(BlockInit.SMELTER_SCREEN_HANDLER, SmelterScreen::new);